(feature) committing the Solution for the first time.
I wrote this program in C# because when I copied several visual studio projects, which are bound to Source Control, the
projects remain bound to TFS and I would get messages when opening them.
I was not able to find an easy way to unbind all the projects, and to unbind them completely you have to modify the .sln files to remove the TFS references.
This program  traverses through a directory looking for any file you specify, and removes a section based on the opening and closing arguments you supply.
Here is an example of how to call it to remove TFS bindings from .sln files:
-d "C:\Users\torre\source\Test" -p"*.sln" -o"TeamFoundationVersionControl" -c"EndGlobalSection"
I made it all configurable so that it can be used for other tasks. I also modify the files in parallel, so that its fast.